external network access from undrecloud
external network access from undrecloud is needed to access
overcloud resources. We need to identify physical interface
which we added to br-ex in controller nodes.

We follow this strategy
1) When controller has seperate nic for external network,
we use the last available interface for external network.
four-nics-vlans config files use this approach.
2) When external network and control plane network share
the same physical nic, we use control plane interface as
external network interface.
three-nics-vlans use this approach.
After deploying overcloud we create a vlan interface on
this external interface and assign ip address to it.