Suppport fake nova driver
This fake driver will create namespace and ports inside
this namespace for our testing in place of VMs.
This allows us avoiding the burden of creating real VMs
when we want to test only network functionality.

To use this driver, copy it to same location of nova fake.py
driver and then update nova.conf with
compute_driver = fake_vif.OVSFakeDriver
and restart nova compute service.

Later we will enhance jetpack to build nova image with this
fake driver.

note: This driver is tested with only osp16 for now.