revert 1821dc96a3736eecd424194c2873966cf39183fa
I requested masco to add commit [1] as I thought
infrared is setting all the nodes to available after
along with introspection command. But looks like it
not doing that and all of the nodes are stuck in
manageble state. So we need to revert this commit [1]
to change their state to available.

[1] 1821dc96a3736eecd424194c2873966cf39183fa